requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-bigquery>=3.17.0
google-cloud-aiplatform>=1.38.0
pydantic>=2.0.0
python-dotenv>=1.0.0
ijson>=3.2.0
requests>=2.31.0
google-cloud-secret-manager>=2.16.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
